begin person refactor
